Seller Dashboard

+++ HireFire .NET/HireFire/Hirefire.Core.Services.Interfaces/ITransactionService.cs
@@ -27,7 +27,7 @@ namespace Hirefire.Core.Services.Interfaces
 
         int TotalIncome(string sellerUserName);
 
-        int LastMonthIncome(string sellerUserName);
+        float LastMonthIncome(string sellerUserName);
